pyproject.toml
[tool.poetry]
name = "hello-wordsmith"
version = "0.1.0"
description = "A simple Python package to wrap llama-index RAG CLI over Wordsmith data."
authors = ["Derek Johnston <derek@wordsmith.ai>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/wordsmith-ai/hello-wordsmith"
classifiers = [
    "Intended Audience :: End Users",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8"
]
packages = [{include = "hello_wordsmith"}]


[tool.poetry.dependencies]
python = "3.9.6"
chromadb = "0.6.3"
llama-index-core = "0.12.22"
llama-index-cli = "0.4.1"
llama-index-llms-ollama = "^0.5.2"
llama-index-vector-stores-chroma = "^0.4.1"
llama-index-readers-file = "^0.4.6"
pydantic = "2.10.6"
onnxruntime = "1.15.1"
numpy = "<2.0.0"
llama-index-embeddings-huggingface = "^0.5.2"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
mypy = "1.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hello-wordsmith = "hello_wordsmith.wordsmith:main"
